+++ Algebras/CRel/TypeConstructors/RealsTC.cpp
@@ -100,13 +100,13 @@ void RealsTC::Delete(const ListExpr typeExpr, Word &value)
   return DefaultDelete(typeExpr, value);
 }
 
-bool RealsTC::Open(SmiRecord &valueRecord, uint64_t &offset,
+bool RealsTC::Open(SmiRecord &valueRecord, size_t &offset,
                   const ListExpr typeExpr, Word &value)
 {
   return DefaultOpen(valueRecord, offset, typeExpr, value);
 }
 
-bool RealsTC::Save(SmiRecord &valueRecord, uint64_t &offset,
+bool RealsTC::Save(SmiRecord &valueRecord, size_t &offset,
                   const ListExpr typeExpr, Word &value)
 {
   return DefaultSave(valueRecord, offset, typeExpr, value);
